+++ app/src/main/java/be/ugent/zeus/hydra/common/request/Request.java
@@ -1,7 +1,7 @@
 package be.ugent.zeus.hydra.common.request;
 
 import android.os.Bundle;
-import androidx.annotation.AnyThread;
+
 import androidx.annotation.NonNull;
 
 import java9.util.function.Function;
