+++ app/src/main/java/be/ugent/zeus/hydra/common/ui/PreferenceFragment.java
@@ -1,7 +1,5 @@
 package be.ugent.zeus.hydra.common.ui;
 
-import android.view.View;
-
 import androidx.annotation.NonNull;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
@@ -14,7 +12,7 @@ import androidx.preference.PreferenceFragmentCompat;
 public abstract class PreferenceFragment extends PreferenceFragmentCompat {
 
     @NonNull
-    public <T extends Preference> T requirePreference(String preference) {
+    protected <T extends Preference> T requirePreference(String preference) {
         T result = findPreference(preference);
         if (result == null) {
             throw new NullPointerException("Preference does not exist, or is loaded too early.");
@@ -22,5 +20,4 @@ public abstract class PreferenceFragment extends PreferenceFragmentCompat {
             return result;
         }
     }
-
 }
