+++ app/src/main/java/be/ugent/zeus/hydra/common/ui/WebViewActivity.java
@@ -4,11 +4,10 @@ import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import androidx.annotation.Nullable;
-
 import android.view.View;
 import android.webkit.WebView;
 import android.widget.ProgressBar;
+
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.common.network.InterceptingWebViewClient;
