+++ app/src/main/java/be/ugent/zeus/hydra/resto/sandwich/ecological/EcologicalFragment.java
@@ -3,6 +3,7 @@ package be.ugent.zeus.hydra.resto.sandwich.ecological;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.*;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
@@ -15,8 +16,8 @@ import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.common.arch.observers.AdapterObserver;
 import be.ugent.zeus.hydra.common.arch.observers.PartialErrorObserver;
 import be.ugent.zeus.hydra.common.arch.observers.ProgressObserver;
-import be.ugent.zeus.hydra.common.ui.BaseActivity;
 import com.google.android.material.snackbar.Snackbar;
+
 import static be.ugent.zeus.hydra.utils.FragmentUtils.requireBaseActivity;
 
 /**
