+++ app/src/test/java/be/ugent/zeus/hydra/association/UgentNewsArticleTest.java
@@ -3,7 +3,6 @@ package be.ugent.zeus.hydra.association;
 import be.ugent.zeus.hydra.association.news.UgentNewsArticle;
 import be.ugent.zeus.hydra.common.ModelTest;
 import be.ugent.zeus.hydra.testing.Utils;
-import nl.jqno.equalsverifier.Warning;
 import org.junit.Test;
 
 /**
