+++ app/src/test/java/be/ugent/zeus/hydra/library/list/LibraryListRequestTest.java
@@ -4,7 +4,6 @@ import be.ugent.zeus.hydra.common.network.AbstractJsonRequestTest;
 import be.ugent.zeus.hydra.common.network.JsonOkHttpRequest;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.RuntimeEnvironment;
 
 /**
  * @author Niko Strijbol
