+++ app/src/test/java/be/ugent/zeus/hydra/resto/extrafood/FoodTest.java
@@ -1,10 +1,6 @@
 package be.ugent.zeus.hydra.resto.extrafood;
 
 import be.ugent.zeus.hydra.common.ModelTest;
-import be.ugent.zeus.hydra.resto.extrafood.Food;
-import nl.jqno.equalsverifier.EqualsVerifier;
-import nl.jqno.equalsverifier.Warning;
-import org.junit.Test;
 
 /**
  * @author Niko Strijbol
