+++ app/src/test/java/be/ugent/zeus/hydra/resto/meta/RestoTest.java
@@ -1,9 +1,6 @@
 package be.ugent.zeus.hydra.resto.meta;
 
 import be.ugent.zeus.hydra.common.ModelTest;
-import nl.jqno.equalsverifier.EqualsVerifier;
-import nl.jqno.equalsverifier.Warning;
-import org.junit.Test;
 
 /**
  * @author Niko Strijbol
