+++ app/src/test/java/be/ugent/zeus/hydra/schamper/ArticleTest.java
@@ -2,7 +2,6 @@ package be.ugent.zeus.hydra.schamper;
 
 import be.ugent.zeus.hydra.common.ModelTest;
 import be.ugent.zeus.hydra.testing.Utils;
-import nl.jqno.equalsverifier.Warning;
 import org.junit.Test;
 
 /**
