+++ app/src/test/java/be/ugent/zeus/hydra/sko/lineup/ArtistTest.java
@@ -1,9 +1,7 @@
 package be.ugent.zeus.hydra.sko.lineup;
 
 import be.ugent.zeus.hydra.common.ModelTest;
-import be.ugent.zeus.hydra.sko.lineup.Artist;
 import be.ugent.zeus.hydra.testing.Utils;
-import nl.jqno.equalsverifier.Warning;
 import org.junit.Test;
 
 /**
